# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a mock HTTP server inspired by httpbin, built with Go and Gorilla Mux.",
        "title": "MockHTTP API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/delay/{seconds}": {
            "get": {
                "description": "Returns a response after a specified delay in seconds",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "delay"
                ],
                "summary": "Delayed response",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delay in seconds",
                        "name": "seconds",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response after {seconds} second(s)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid delay time",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/status/{code}": {
            "get": {
                "description": "Returns a response with the specified status code",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Return custom status code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid status code",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}